Chapter05/myblog/requirements.txt
Django
djangorestframework
psycopg2-binary
python-social-auth
